Saving

+++ DiceGuess.cpp
@@ -33,7 +33,7 @@ void Player::showDices()
     cout << "{ ";
     for (int i = 0; i != DiceNumber; ++i)
     {
-        cout << dices[i] << ' ';
+        cout << static_cast<int>(dices[i]) << ' ';
     }
     cout << "}\n";
 }
